makeSummarizedExperimentFromLoom {SummarizedExperiment}	R Documentation

Make a SummarizedExperiment from a '.loom' hdf5 file

Description

makeSummarizedExperimentFromLoom represents a '.loom' file as a SummarizedExperiment. The '/matrix' and '/layers' are represented as HDF5Array objects; row and column attributes are parsed to DataFrame. Optionally, row or column attributes can be specified as row and and column names.

Usage

makeSummarizedExperimentFromLoom(file,
                                 rownames_attr = NULL,
                                 colnames_attr = NULL)

Arguments

file
The path (as a single character string) to the HDF5 file where the dataset is located.

rownames_attr
The name of the row attribute to be used as row names.

colnames_attr
The name of the column attribute to be used as column names.

Value

A SummarizedExperiment object with row and column data and one or more assays.

Author(s)

Martin Morgan

See Also

http://loompy.org/loompy-docs/format/index.html for a specification of the .loom format.

Examples

## ---------------------------------------------------------------------
## BASIC EXAMPLE
## ---------------------------------------------------------------------

file <- system.file(
    package="SummarizedExperiment", "extdata", "example.loom"
)
se <- makeSummarizedExperimentFromLoom(file)
se
assay(se)
metadata(se)

[Package SummarizedExperiment version 1.20.0 Index]